\name{DisulfRisk-package}
\alias{DisulfRisk-package}
\alias{DisulfRisk}
\title{DisulfRisk: Weighted DRS Scoring for Expression Cohorts}
\description{
Tools for reading bulk expression matrices and calculating a
disulfidptosis-related weighted DRS score using a fixed 20-gene model.
}
\keyword{internal}
